requirements_dev.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
lxml==4.9.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-http-client==3.3.7
requests==2.28.1
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
urllib3==1.26.12
wget==3.2
XlsxWriter==3.0.3
autopep8==1.7.0
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
